refactor: move util funcs to utils.rs
