+++ src/lib.rs
@@ -4,11 +4,12 @@ use once_cell::sync::Lazy;
 mod load_res;
 mod models;
 mod requests;
+mod utils;
 
 pub use models::api::{Status, WarInfo, PlanetStatus, PlanetAttack, Campaign, GlobalEvent, HomeWorld, Position, PlanetInfo};
 pub use models::{Planet, Faction, Sector, Language};
 pub use requests::{get_status, get_war_info};
-
+pub use utils::{get_total_player_count, get_top_planets_by_player_count, get_faction_distribution};
 
 /// The base URL for the Helldivers API
 pub const BASE_URL: &str = "https://api.live.prod.thehelldiversgame.com/api";
@@ -47,36 +48,6 @@ pub fn get_sector_name(id: i64) -> Option<String> {
     SECTORS.get(&id).map(|sector| sector.name.clone())
 }
 
-/// Get the total player count for a status
-/// 
-/// Arguments:
-///  status: Status - The status to get the player count for
-pub fn get_total_player_count(status: &Status) -> i64 {
-    status.planet_status.iter().map(|ps| ps.players).sum()
-}
-
-/// Get the top planets by player count
-/// 
-/// Arguments:
-///  status: &Status - The status to get the top planets from
-///  count: usize - The number of top planets to get
-pub fn get_top_planets_by_player_count(status: &Status, count: usize) -> Vec<(&PlanetStatus, i64)> {
-    let mut planet_players: Vec<(&PlanetStatus, i64)> = status.planet_status.iter().map(|ps| (ps, ps.players)).collect();
-    planet_players.sort_by(|a, b| b.1.cmp(&a.1));
-    planet_players.into_iter().take(count).collect()
-}
-
-/// Get the faction distribution 
-/// 
-/// Arguments:
-///  status: &Status - The status to get the faction distribution from
-pub fn get_faction_distribution(status: &Status) -> HashMap<i64, i64> {
-    let mut distribution = HashMap::new();
-    for ps in &status.planet_status {
-        *distribution.entry(ps.owner).or_insert(0) += 1;
-    }
-    distribution
-}
 #[cfg(test)]
 mod tests {
     use super::*;
